+++ src/app/services/groceries.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Item } from './models/item';
+import { Item } from '../models/item';
 import { HttpClient } from '@angular/common/http';
 import { BehaviorSubject, firstValueFrom } from 'rxjs';
 import { environment } from 'src/environments/environment';
